backend/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
h11~=0.14.0
jupyter~=1.0.0
pip~=22.0.4
attrs~=21.4.0
wheel~=0.40.0
Pillow~=9.4.0
tornado~=6.2
anyio~=3.6.2
sniffio~=1.3.0
colorama~=0.4.5
python-dotenv~=1.0.0
blinker~=1.6.2
fastapi~=0.95.0
starlette~=0.26.1
pydantic~=1.10.7
requests~=2.28.2
uvicorn~=0.21.1
decorator~=5.1.1
greenlet~=2.0.2
PyYAML~=6.0
setuptools~=58.1.0
pymongo~=4.4.1
beautifulsoup4~=4.12.0
soupsieve~=2.4
dnspython~=2.4.2
idna~=3.3
defusedxml~=0.7.1
cffi~=1.15.1
ipython~=8.4.0
pyzmq~=25.0.2
traitlets~=5.9.0
fqdn~=1.5.1
parso~=0.8.3
jedi~=0.18.1
QtPy~=2.3.1
packaging~=21.3
arrow~=1.2.3
python-dateutil~=2.8.2
pywin32~=304
webencodings~=0.5.1
bleach~=6.0.0
tinycss2~=1.2.1
psutil~=5.9.1
certifi~=2021.10.8
debugpy~=1.6.2
Django~=4.1.7
regex~=2023.6.3
Pygments~=2.12.0
numpy~=1.23.5
matplotlib~=3.7.1
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.5
asgiref~=3.6.0
nbformat~=5.8.0
nbclient~=0.7.2
nbconvert~=7.2.10
fastjsonschema~=2.16.3
jsonschema~=4.17.3
notebook~=6.5.3
Send2Trash~=1.8.0
terminado~=0.17.1
nbclassic~=0.5.3
urllib3~=1.26.15
asttokens~=2.0.5
executing~=0.8.3
ipykernel~=6.15.1
Markdown~=3.4.3
pandocfilters~=1.5.0
mistune~=2.0.5
entrypoints~=0.4
pycparser~=2.21
qtconsole~=5.4.1
websocket-client~=1.5.1
ipywidgets~=8.0.6
pyrsistent~=0.19.3
webcolors~=1.13
jsonpointer~=2.3
isoduration~=20.11.0
platformdirs~=3.10.0